move remove mutation to own file
